requirements.txt
# Conda is used to install ffmpeg
# conda install -c conda-forge ffmpeg
uvicorn==0.23.1 
fastapi==0.100.0
numpy==1.24.4
pydub==0.25.1
requests==2.31.0
soundfile==0.12.1 
pyyaml==6.0
python-multipart==0.0.6